Update Hook to add "useDebouncedCallback"
